+++ github-updater.php
@@ -12,7 +12,7 @@
 Plugin Name:       GitHub Updater
 Plugin URI:        https://github.com/afragen/github-updater
 Description:       A plugin to automatically update GitHub or Bitbucket hosted plugins and themes into WordPress. Plugin class based upon <a href="https://github.com/codepress/github-plugin-updater">codepress/github-plugin-updater</a>. Theme class based upon <a href="https://github.com/WordPress-Phoenix/whitelabel-framework">Whitelabel Framework</a> modifications.
-Version:           2.9.0
+Version:           3.0.0
 Author:            Andy Fragen
 License:           GNU General Public License v2
 License URI:       http://www.gnu.org/licenses/gpl-2.0.html
@@ -29,7 +29,8 @@ if ( ! defined( 'WPINC' ) ) {
 
 // Load main class
 if ( ! class_exists( 'GitHub_Updater' ) ) {
-	require_once 'includes/class-github-updater.php';
+	require 'includes/class-github-updater.php';
+	load_plugin_textdomain( 'github-updater', false, dirname( plugin_basename( __FILE__ ) ) . '/languages' );
 }
 
 // Instantiate class GitHub_Updater