+++ mu/ghu-loader.php
@@ -12,7 +12,7 @@
 Plugin Name:       GitHub Updater MU loader
 Plugin URI:        https://github.com/afragen/github-updater
 Description:       A plugin to automatically update GitHub or Bitbucket hosted plugins and themes into WordPress. Disables normal plugin activation and deletion.
-Version:           1.1.0
+Version:           1.2.0
 Author:            Andy Fragen
 License:           GNU General Public License v2
 License URI:       http://www.gnu.org/licenses/gpl-2.0.html
@@ -26,13 +26,14 @@ if ( ! defined( 'WPINC' ) ) {
 
 // Load normal plugin
 if ( ! class_exists( 'GitHub_Updater' ) ) {
-	define( 'GHU_PLUGIN_FILE', 'github-updater/github-updater.php' );
-	require_once trailingslashit( WP_PLUGIN_DIR ). GHU_PLUGIN_FILE;
+	$ghu_plugin_file = 'github-updater/github-updater.php';
+	require trailingslashit( WP_PLUGIN_DIR ). $ghu_plugin_file;
 }
 
 function ghu_deactivate( $plugin, $network_wide ) {
-	if ( GHU_PLUGIN_FILE === $plugin ) {
-		deactivate_plugins( GHU_PLUGIN_FILE );
+	$ghu_plugin_file = 'github-updater/github-updater.php';
+	if ( $ghu_plugin_file === $plugin ) {
+		deactivate_plugins( $ghu_plugin_file );
 	}
 }
 
@@ -43,13 +44,16 @@ function ghu_mu_plugin_active( $actions ) {
 	if ( isset( $actions['delete'] ) ) {
 		unset( $actions['delete'] );
 	}
+	if ( isset( $actions['deactivate'] ) ) {
+		unset( $actions['deactivate'] );
+	}
 
-	return array_merge( array( 'mu-plugin' => 'Activated as mu-plugin' ), $actions );
+	return array_merge( array( 'mu-plugin' => __('Activated as mu-plugin', 'github-updater' ) ), $actions );
 }
 
 // Deactivate normal plugin as it's loaded as mu-plugin
 add_action( 'activated_plugin', 'ghu_deactivate', 10, 2 );
 
 // Remove links from Plugins page so user can't delete main plugin
-add_filter( 'network_admin_plugin_action_links_' . GHU_PLUGIN_FILE, 'ghu_mu_plugin_active' );
-add_filter( 'plugin_action_links_' . GHU_PLUGIN_FILE, 'ghu_mu_plugin_active' );
+add_filter( 'network_admin_plugin_action_links_' . $ghu_plugin_file, 'ghu_mu_plugin_active' );
+add_filter( 'plugin_action_links_' . $ghu_plugin_file, 'ghu_mu_plugin_active' );